requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
babel==2.16.0
bibtexparser==2.0.0b8
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
clldutils==3.24.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
csvw==3.5.1
dlinfo==1.2.1
espeakng-loader==0.2.2
exceptiongroup==1.2.2
filelock==3.16.1
flatbuffers==24.12.23
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
language-tags==1.2.0
lightning-whisper-mlx==0.0.10
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
mlx
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
ollama==0.4.6
onnxruntime==1.19.2
packaging==24.2
phonemizer-fork @ git+https://github.com/thewh1teagle/phonemizer-fork@aa26b4cf2c518386a8c2dcea5c6f0866cde10bd3
protobuf==5.29.3
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pylatexenc==2.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
rdflib==7.1.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.22.3
scipy==1.13.1
segments==2.2.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
sympy==1.13.1
tabulate==0.9.0
tiktoken==0.3.3
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
zipp==3.21.0
openai-whisper
torch
torchaudio
ctranslate2>=4.5.0
faster-whisper
python-dotenv
